[PATRO23-41] change location of event handlers to Budget.Application
