+++ src/modules/budget/application/EventHandlers/BudgetTransactionSoftDeleteDomainEventHandler.cs
@@ -4,21 +4,21 @@ using Intive.Patronage2023.Shared.Infrastructure.EventHandlers;
 
 using Microsoft.Extensions.DependencyInjection;
 
-namespace Intive.Patronage2023.Modules.Budget.Infrastructure.Domain.EventHandlers;
+namespace Intive.Patronage2023.Modules.Budget.Application.EventHandlers;
 
 /// <summary>
-/// Budget Transaction created domain event handler.
+/// This class is a domain event handler that handles the "BudgetTransactionSoftDeleteDomainEvent" event.
 /// </summary>
 [Lifetime(Lifetime = ServiceLifetime.Singleton)]
-public class BudgetTransactionCreatedDomainEventHandler : IDomainEventHandler<BudgetTransactionCreatedDomainEvent>
+public class BudgetTransactionSoftDeleteDomainEventHandler : IDomainEventHandler<BudgetTransactionSoftDeleteDomainEvent>
 {
 	/// <summary>
-	/// Handle the notification.
+	/// This method is the entry point for handling the "BudgetTransactionSoftDeleteDomainEvent" event.
 	/// </summary>
 	/// <param name="notification">Notification.</param>
 	/// <param name="cancellationToken">Cancelation token.</param>
 	/// <returns>Task.</returns>
-	public Task Handle(BudgetTransactionCreatedDomainEvent notification, CancellationToken cancellationToken)
+	public Task Handle(BudgetTransactionSoftDeleteDomainEvent notification, CancellationToken cancellationToken)
 	{
 		cancellationToken.ThrowIfCancellationRequested();
